fix: resolve code smells and duplication
