+++ packages/web-service/src/pages/site/select-address/select-address.js
@@ -2,9 +2,9 @@ import Joi from 'joi'
 import pageRoute from '../../../routes/page-route.js'
 import { APIRequests } from '../../../services/api-requests.js'
 import { siteURIs } from '../../../uris.js'
-import { isCompleteOrConfirmed } from '../../common/tag-functions.js'
+import { checkApplication } from '../../common/check-application.js'
 import { mapLookedUpAddress } from '../../contact/common/address/address.js'
-import { SECTION_TASKS } from '../../tasklist/licence-type-map.js'
+import { siteAddressCompletion, getSite } from '../common/site-map-upload.js'
 
 export const getData = async request => {
   const journeyData = await request.cache().getData()
@@ -23,31 +23,19 @@ export const setData = async request => {
   const { Address: lookupAddress } = addressLookup.find(a => Number.parseInt(a.Address.UPRN) === request.payload.uprn)
 
   const apiAddress = mapLookedUpAddress(lookupAddress)
-  const site = await APIRequests.SITE.findByApplicationId(applicationId)
-  let siteInfo = {}
-  if (site.length) {
-    siteInfo = site[0]
-  }
+  const siteInfo = await getSite(applicationId)
   const payload = { ...siteInfo, address: apiAddress }
   await APIRequests.SITE.update(siteInfo.id, payload)
   journeyData.siteData = { ...siteData, address: apiAddress }
   await request.cache().setData(journeyData)
 }
 
-export const completion = async request => {
-  const { applicationId } = await request.cache().getData()
-
-  const appTagStatus = await APIRequests.APPLICATION.tags(applicationId).get(SECTION_TASKS.SITES)
-  if (isCompleteOrConfirmed(appTagStatus)) {
-    return siteURIs.CHECK_SITE_ANSWERS.uri
-  }
-
-  return siteURIs.UPLOAD_MAP.uri
-}
+export const completion = async request => await siteAddressCompletion(request)
 
 export default pageRoute({
   page: siteURIs.SELECT_ADDRESS.page,
   uri: siteURIs.SELECT_ADDRESS.uri,
+  checkData: checkApplication,
   validator: Joi.object({
     uprn: Joi.number()
   }).options({ abortEarly: false, allowUnknown: true }),
